SwiperEventos
la implementacion de swiper a eventos a sido un exito .... recomiendo borrar el usehook que ya no se usara amen
