+++ src/components/SliderEventos/CardEventos.Styled.js
@@ -6,7 +6,7 @@ padding: 20px;
   background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)),
     url('https://media.30seconds.com/tip/lg/Want-to-Run-a-Marathon-Why-You-Need-to-Talk-to-Your-Doctor-7273-1a98fb04bb-1502290274.jpg')
       center center/cover no-repeat;
-  width: 340px;
+  width: 400px;
   height: 300px;
   color: white;
   display: flex;
